Update profile page link to find circles
